milvus-service/requirements.txt
pymilvus==2.5.3
sentence-transformers==3.3.1
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.32.1
jinja2==3.1.4
python-multipart==0.0.20

psycopg[binary,pool]==3.2.13
requests==2.32.3
